pyproject.toml
[tool.poetry]
name = "notificationbot"
version = "0.1.0"
description = ""
authors = ["blacksmokezip <hanzomainer1488@gmail.com>"]
readme = "README.md"
packages = [{include = "bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.2.0"
asyncio = "^3.4.3"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
